catalog_project/requirements.txt
asgiref==3.8.1
Django==5.1.7
django-filter==25.1
django-grappelli==4.0.1
django-redis==5.4.0
djangorestframework==3.16.0
drf-extensions==0.7.1
pillow==11.1.0
psycopg2-binary==2.9.10
redis==5.2.1
sqlparse==0.5.3
tzdata==2025.2